Rename variant analysis data and domain mappers to follow convention
