+++ extensions/ql-vscode/src/query-history/store/query-history-variant-analysis-dto-mapper.ts
